Move TextureManager into managers folder
