+++ src/managers/TextureManager.ts
